Clean common physics code up
