+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "crates/*",
-    "examples/common",
+    "examples/common*",
     "examples/bevy_gltf_components/*",
     "examples/bevy_gltf_blueprints/*",
     "examples/bevy_gltf_save_load/*",
