+++ examples/bevy_gltf_blueprints/animation/src/game/in_game.rs
@@ -1,4 +1,6 @@
-use bevy_gltf_worlflow_examples_common::{assets::GameAssets, GameState, InAppRunning, Player};
+use bevy_gltf_worlflow_examples_common_rapier::{
+    assets::GameAssets, GameState, InAppRunning, Player,
+};
 use bevy_rapier3d::prelude::Velocity;
 use rand::Rng;
 use std::time::Duration;
