+++ examples/bevy_gltf_blueprints/basic/Cargo.toml
@@ -7,6 +7,6 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 bevy = { version = "0.13", features = ["dynamic_linking"] }
 bevy_gltf_blueprints = { path = "../../../crates/bevy_gltf_blueprints" }
-bevy_gltf_worlflow_examples_common = { path = "../../common" }
+bevy_gltf_worlflow_examples_common_rapier = { path = "../../common_rapier" }
 bevy_rapier3d = { version = "0.25.0", features = ["serde-serialize", "debug-render-3d", "enhanced-determinism"] }
 rand = "0.8.5"
