+++ examples/bevy_gltf_blueprints/basic/src/game/in_main_menu.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use bevy_gltf_worlflow_examples_common::{AppState, InMainMenu};
+use bevy_gltf_worlflow_examples_common_rapier::{AppState, InMainMenu};
 
 pub fn setup_main_menu(mut commands: Commands) {
     commands.spawn((
