+++ examples/bevy_gltf_blueprints/basic/src/main.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use bevy_gltf_worlflow_examples_common::CommonPlugin;
+use bevy_gltf_worlflow_examples_common_rapier::CommonPlugin;
 
 mod core;
 use crate::core::*;
