+++ examples/bevy_gltf_blueprints/basic_xpbd_physics/Cargo.toml
@@ -7,6 +7,6 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 bevy = { version = "0.13", features = ["dynamic_linking"] }
 bevy_gltf_blueprints = { path = "../../../crates/bevy_gltf_blueprints" }
-bevy_gltf_worlflow_examples_common = { path = "../../common", default-features = false, features = ["blueprints", "physics_xpbd"] }
+bevy_gltf_worlflow_examples_common_xpbd = { path = "../../common_xpbd" }
 bevy_xpbd_3d = "0.4"
 rand = "0.8.5"
