+++ examples/bevy_gltf_blueprints/basic_xpbd_physics/src/game/in_game.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 use bevy_gltf_blueprints::{BluePrintBundle, BlueprintName, GameWorldTag};
-use bevy_gltf_worlflow_examples_common::{assets::GameAssets, GameState, InAppRunning};
+use bevy_gltf_worlflow_examples_common_xpbd::{assets::GameAssets, GameState, InAppRunning};
 use bevy_xpbd_3d::prelude::*;
 use rand::Rng;
