+++ examples/bevy_gltf_blueprints/multiple_levels_multiple_blendfiles/src/game/mod.rs
@@ -8,7 +8,7 @@ pub mod level_transitions;
 pub use level_transitions::*;
 
 use bevy::prelude::*;
-use bevy_gltf_worlflow_examples_common::{AppState, GameState};
+use bevy_gltf_worlflow_examples_common_rapier::{AppState, GameState};
 
 pub struct GamePlugin;
 impl Plugin for GamePlugin {
