+++ examples/bevy_gltf_components/basic/Cargo.toml
@@ -7,4 +7,4 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 bevy = { version = "0.13", features = ["dynamic_linking"] }
 bevy_gltf_components = { path = "../../../crates/bevy_gltf_components" }
-bevy_gltf_worlflow_examples_common = { path = "../../common" }
+bevy_gltf_worlflow_examples_common_rapier = { path = "../../common_rapier" }
