+++ examples/bevy_gltf_save_load/basic/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 bevy = { version = "0.13", features = ["dynamic_linking"] }
 bevy_gltf_blueprints = { path = "../../../crates/bevy_gltf_blueprints" }
 bevy_gltf_save_load = { path = "../../../crates/bevy_gltf_save_load" }
-bevy_gltf_worlflow_examples_common = { path = "../../common" }
+bevy_gltf_worlflow_examples_common_rapier = { path = "../../common_rapier" }
 
 serde_json = "1.0.108"
 serde = "1.0.193"
