+++ examples/bevy_gltf_save_load/basic/src/game/in_game.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use bevy_gltf_blueprints::{BluePrintBundle, BlueprintName, GameWorldTag, Library, NoInBlueprint};
 use bevy_gltf_save_load::{Dynamic, DynamicEntitiesRoot, StaticEntitiesRoot};
-use bevy_gltf_worlflow_examples_common::{GameState, InAppRunning, Player};
+use bevy_gltf_worlflow_examples_common_rapier::{GameState, InAppRunning, Player};
 use bevy_rapier3d::prelude::Velocity;
 use rand::Rng;
