+++ examples/common_rapier/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bevy_gltf_worlflow_examples_common_rapier"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[features]
+blueprints = ["dep:bevy_gltf_blueprints"]
+default = ["blueprints"]
+
+
+[dependencies]
+bevy = { version = "0.13", features = ["dynamic_linking"] }
+bevy_gltf_worlflow_examples_common = { path = "../common" }
+bevy_gltf_blueprints = { path = "../../crates/bevy_gltf_blueprints", optional = true }
+bevy_rapier3d = { version = "0.25", features = ["serde-serialize", "debug-render-3d", "enhanced-determinism"] }
+bevy_asset_loader = { version = "0.20", features = ["standard_dynamic_assets"] }
+bevy_editor_pls = { version = "0.8" }
+rand = "0.8.5"
