+++ examples/common_rapier/src/physics/physics_replace_proxies.rs
