+++ examples/common_rapier/src/physics/utils.rs
