+++ examples/common_xpbd/src/physics/mod.rs
@@ -6,21 +6,14 @@ pub(crate) mod utils;
 pub(crate) mod controls;
 pub(crate) use controls::*;
 
-use crate::state::GameState;
 use bevy::prelude::*;
-use bevy_gltf_blueprints::GltfBlueprintsSet;
-use bevy_rapier3d::{
-    prelude::{NoUserData, RapierPhysicsPlugin},
-    render::RapierDebugRenderPlugin,
-};
+use bevy_xpbd_3d::prelude::*;
 
-pub struct PhysicsPlugin;
-impl Plugin for PhysicsPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_plugins((
-            RapierPhysicsPlugin::<NoUserData>::default(),
-            RapierDebugRenderPlugin::default(),
-        ))
+use bevy_gltf_blueprints::GltfBlueprintsSet;
+use bevy_gltf_worlflow_examples_common::state::GameState;
+
+pub(crate) fn plugin(app: &mut App) {
+    app.add_plugins((PhysicsPlugins::default(), PhysicsDebugPlugin::default()))
         .register_type::<AutoAABBCollider>()
         .register_type::<physics_replace_proxies::Collider>()
         .add_systems(
@@ -30,5 +23,4 @@ impl Plugin for PhysicsPlugin {
         .add_systems(Update, toggle_physics_debug)
         .add_systems(OnEnter(GameState::InGame), resume_physics)
         .add_systems(OnExit(GameState::InGame), pause_physics);
-    }
 }
