+++ examples/common_xpbd/src/physics/physics_replace_proxies.rs
