+++ examples/common_xpbd/src/physics/utils.rs
