+++ testing/bevy_example/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 bevy = { version = "0.13", features = ["dynamic_linking"] }
 bevy_gltf_blueprints = { path = "../../crates/bevy_gltf_blueprints" }
 bevy_registry_export = { path = "../../crates/bevy_registry_export" }
-bevy_gltf_worlflow_examples_common = { path = "../../examples/common" }
+bevy_gltf_worlflow_examples_common_rapier = { path = "../../examples/common_rapier" }
 
 bevy_rapier3d = { version = "0.25.0", features = ["serde-serialize", "debug-render-3d", "enhanced-determinism"] }
 bevy_asset_loader = { version = "0.20", features = ["standard_dynamic_assets"] }
